requirements.txt
streamlit
python-dotenv
langchain-community
langchain-core
langchain-groq
pypdf
python-docx
sentence-transformers
faiss-cpu
